+++ packages/avatar/Avatar.vue
@@ -21,6 +21,7 @@ export default {
   props: {
     size: {
       type: [Number, String],
+      default: 'large',
       validator(val) {
         if (typeof val === 'string') {
           return ['large', 'medium', 'small'].includes(val)
@@ -51,9 +52,9 @@ export default {
 
     const isImageExist = ref(true)
 
-    const handleError = () => {
-      const errorFlag = error ? error.value() : undefined
-      if (errorFlag !== false) {
+    const handleError = (e) => {
+      const errorFlag = error?.value(e)
+      if (!!errorFlag !== false) {
         isImageExist.value = false
       }
     }