+++ packages/date-picker/src/panel/date-range.vue
@@ -237,13 +237,13 @@ import {
   nextDate,
   extractDateFormat,
   extractTimeFormat
-} from 'element-ui/src/utils/date-util'
-import Clickoutside from 'element-ui/src/utils/clickoutside'
-import Locale from 'element-ui/src/mixins/locale'
+} from '../../../../src/utils/date-util'
+import Clickoutside from '../../../../src/utils/clickoutside'
+import Locale from '../../../../src/mixins/locale'
 import TimePicker from './time'
 import DateTable from '../basic/date-table'
-import ElInput from 'element-ui/packages/input'
-import ElButton from 'element-ui/packages/button'
+import ElInput from '../../../input'
+import ElButton from '../../../button'
 
 const calcDefaultValue = (defaultValue) => {
   if (Array.isArray(defaultValue)) {