+++ packages/message-box/index.js
@@ -5,14 +5,4 @@ export function useMsgbox() {
   return MessageBox
 }
 
-export function useAlert() {
-  return MessageBox.alert
-}
-export function useConfirm() {
-  return MessageBox.confirm
-}
-export function usePrompt() {
-  return MessageBox.prompt
-}
-
 export default MessageBoxComponent