+++ packages/table/src/filter-panel.vue
@@ -64,14 +64,14 @@
 </template>
 
 <script type="text/babel">
-import Popper from 'element-ui/src/utils/vue-popper'
-import { PopupManager } from 'element-ui/src/utils/popup'
-import Locale from 'element-ui/src/mixins/locale'
-import Clickoutside from 'element-ui/src/utils/clickoutside'
+import Popper from '../../../src/utils/vue-popper'
+import { PopupManager } from '../../../src/utils/popup'
+import Locale from '../../../src/mixins/locale'
+import Clickoutside from '../../../src/utils/clickoutside'
 import Dropdown from './dropdown'
-import ElCheckbox from 'element-ui/packages/checkbox'
-import ElCheckboxGroup from 'element-ui/packages/checkbox-group'
-import ElScrollbar from 'element-ui/packages/scrollbar'
+import ElCheckbox from '../../checkbox'
+import ElCheckboxGroup from '../../checkbox-group'
+import ElScrollbar from '../../scrollbar'
 
 export default {
   name: 'ElTableFilterPanel',