+++ packages/table/src/table-header.js
@@ -1,7 +1,7 @@
 // import {h} from 'vue'
 import * as Vue from 'vue'
-import { hasClass, addClass, removeClass } from 'element-ui/src/utils/dom'
-import ElCheckbox from 'element-ui/packages/checkbox'
+import { hasClass, addClass, removeClass } from '../../../src/utils/dom'
+import ElCheckbox from '../../checkbox'
 import FilterPanel from './filter-panel.vue'
 import LayoutObserver from './layout-observer'
 import { mapStates } from './store/helper'
@@ -245,7 +245,7 @@ export default {
     })
   },
 
-  beforeDestroy() {
+  beforeUnmount() {
     const panels = this.filterPanels
     for (const prop in panels) {
       if (Object.hasOwnProperty.call(panels, prop) && panels[prop]) {