+++ types/element3.d.ts
@@ -1,64 +1,76 @@
-import Vue, { PluginObject } from 'vue'
-import { ElementUIComponent, ElementUIComponentSize, ElementUIHorizontalAlignment } from './component'
+import { PluginObject, App } from 'vue'
+import {
+  ElementUIComponent,
+  ElementUIComponentSize,
+  ElementUIHorizontalAlignment
+} from './component'
 
-import { ElAlert } from './alert'
-import { ElAside } from './aside'
-import { ElAutocomplete } from './autocomplete'
-import { ElBadge } from './badge'
-import { ElBreadcrumb } from './breadcrumb'
-import { ElBreadcrumbItem } from './breadcrumb-item'
-import { ElButton } from './button'
-import { ElButtonGroup } from './button-group'
-import { ElCard } from './card'
-import { ElCarousel } from './carousel'
-import { ElCarouselItem } from './carousel-item'
-import { ElCascader } from './cascader'
-import { ElCheckbox } from './checkbox'
-import { ElCheckboxButton } from './checkbox-button'
-import { ElCheckboxGroup } from './checkbox-group'
-import { ElCol } from './col'
+export { ElAlert } from './alert'
+export { ElAside } from './aside'
+export { ElAutocomplete } from './autocomplete'
+export { ElBadge } from './badge'
+export { ElBreadcrumb } from './breadcrumb'
+export { ElBreadcrumbItem } from './breadcrumb-item'
+export { ElButton } from './button'
+export { ElButtonGroup } from './button-group'
+export { ElCard } from './card'
+export { ElCarousel } from './carousel'
+export { ElCarouselItem } from './carousel-item'
+export { ElCascader } from './cascader'
+export { ElCheckbox } from './checkbox'
+export { ElCheckboxButton } from './checkbox-button'
+export { ElCheckboxGroup } from './checkbox-group'
+export { ElRow } from './row'
+export { ElCol } from './col'
+export { ElMain } from './main'
+export { ElContainer } from './container'
+export { ElHeader } from './header'
+export { ElFooter } from './footer'
+export { ElProgress } from './progress'
+export { ElPagination } from './pagination'
+export { ElSelect } from './select'
+export { ElRadioButton } from './radio-button'
+export { ElInput } from './input'
+export { ElInputNumber } from './input-number'
+export { ElRadio } from './radio'
+export { ElRadioGroup } from './radio-group'
+export { ElSlider } from './slider'
+export { ElSwitch } from './switch'
+export { ElLink } from './link'
+export { ElAvatar } from './avatar'
+export { ElTag } from './tag'
+export { ElRate } from './rate'
+export { useMessage } from './message'
+export { useNotify } from './notification'
+export { useLoading } from './loading'
+export { useMsgbox } from './message-box'
+
 import { ElCollapse } from './collapse'
 import { ElCollapseItem } from './collapse-item'
 import { ElColorPicker } from './color-picker'
-import { ElContainer } from './container'
 import { ElDatePicker } from './date-picker'
 import { ElDialog } from './dialog'
 import { ElDropdown } from './dropdown'
 import { ElDropdownItem } from './dropdown-item'
 import { ElDropdownMenu } from './dropdown-menu'
-import { ElFooter } from './footer'
 import { ElForm } from './form'
 import { ElFormItem } from './form-item'
-import { ElHeader } from './header'
 import { ElInput } from './input'
 import { ElInputNumber } from './input-number'
 import { ElLoading } from './loading'
-import { ElMain } from './main'
 import { ElMenu } from './menu'
 import { ElMenuItem } from './menu-item'
 import { ElMenuItemGroup } from './menu-item-group'
-import { ElMessage } from './message'
-import { ElMessageBox } from './message-box'
 import { ElNotification } from './notification'
 import { ElOption } from './option'
 import { ElOptionGroup } from './option-group'
-import { ElPagination } from './pagination'
 import { ElPopover } from './popover'
-import { ElProgress } from './progress'
-import { ElRate } from './rate'
-import { ElRadio } from './radio'
-import { ElRadioButton } from './radio-button'
-import { ElRadioGroup } from './radio-group'
-import { ElRow } from './row'
-import { ElSelect } from './select'
-import { ElSlider } from './slider'
+
 import { ElStep } from './step'
 import { ElSteps } from './steps'
 import { ElSubmenu } from './submenu'
-import { ElSwitch } from './switch'
 import { ElTable } from './table'
 import { ElTableColumn } from './table-column'
-import { ElTag } from './tag'
 import { ElTabs } from './tabs'
 import { ElTabPane } from './tab-pane'
 import { ElTimeline } from './timeline'
@@ -69,7 +81,6 @@ import { ElTooltip } from './tooltip'
 import { ElTransfer } from './transfer'
 import { ElTree, TreeData } from './tree'
 import { ElUpload } from './upload'
-import { ElLink } from './link'
 import { ElDivider } from './divider'
 import { ElIcon } from './icon'
 import { ElCalendar } from './calendar'
@@ -77,13 +88,12 @@ import { ElImage } from './image'
 import { ElBacktop } from './backtop'
 import { ElInfiniteScroll } from './infinite-scroll'
 import { ElPageHeader } from './page-header'
-import { ElAvatar } from './avatar'
 import { ElDrawer } from './drawer'
 import { ElPopconfirm } from './popconfirm'
 
 export interface InstallationOptions {
-  locale: any,
-  i18n: any,
+  locale: any
+  i18n: any
   size: string
 }
 
@@ -93,9 +103,9 @@ export const version: string
 /**
  * Install all element-ui components into Vue.
  * Please do not invoke this method directly.
- * Call `Vue.use(ElementUI)` to install.
+ * Call `app.use(ElementUI)` to install.
  */
-export function install (vue: typeof Vue, options: InstallationOptions): void
+export function install(app: App, ...options: any[]): any
 
 /** ElementUI component common definition */
 export type Component = ElementUIComponent
@@ -109,64 +119,8 @@ export type HorizontalAlignment = ElementUIHorizontalAlignment
 /** Show animation while loading data */
 export const Loading: ElLoading
 
-/** Used to show feedback after an activity. The difference with Notification is that the latter is often used to show a system level passive notification. */
-export const Message: ElMessage
-
-/** A set of modal boxes simulating system message box, mainly for message prompt, success tips, error messages and query information */
-export const MessageBox: ElMessageBox
-
 /** Displays a global notification message at the upper right corner of the page */
 export const Notification: ElNotification
-
-// TS cannot merge imported class with namespace, so declare subclasses instead
-
-/** Alert Component */
-export class Alert extends ElAlert {}
-
-/** Aside Component */
-export class Aside extends ElAside {}
-
-/** Autocomplete Component */
-export class Autocomplete extends ElAutocomplete {}
-
-/** Bagde Component */
-export class Badge extends ElBadge {}
-
-/** Breadcrumb Component */
-export class Breadcrumb extends ElBreadcrumb {}
-
-/** Breadcrumb Item Component */
-export class BreadcrumbItem extends ElBreadcrumbItem {}
-
-/** Button Component */
-export class Button extends ElButton {}
-
-/** Button Group Component */
-export class ButtonGroup extends ElButtonGroup {}
-
-/** Card Component */
-export class Card extends ElCard {}
-
-/** Cascader Component */
-export class Cascader extends ElCascader {}
-
-/** Carousel Component */
-export class Carousel extends ElCarousel {}
-
-/** Carousel Item Component */
-export class CarouselItem extends ElCarouselItem {}
-
-/** Checkbox Component */
-export class Checkbox extends ElCheckbox {}
-
-/** Checkbox Button Component */
-export class CheckboxButton extends ElCheckboxButton {}
-
-/** Checkbox Group Component */
-export class CheckboxGroup extends ElCheckboxGroup {}
-
-/** Colunm Layout Component */
-export class Col extends ElCol {}
 
 /** Collapse Component */
 export class Collapse extends ElCollapse {}
@@ -176,9 +130,6 @@ export class CollapseItem extends ElCollapseItem {}
 
 /** Color Picker Component */
 export class ColorPicker extends ElColorPicker {}
-
-/** Container Component */
-export class Container extends ElContainer {}
 
 /** Date Picker Component */
 export class DatePicker extends ElDatePicker {}
@@ -195,26 +146,17 @@ export class DropdownItem extends ElDropdownItem {}
 /** Dropdown Menu Component */
 export class DropdownMenu extends ElDropdownMenu {}
 
-/** Footer Component */
-export class Footer extends ElFooter {}
-
 /** Form Component */
 export class Form extends ElForm {}
 
 /** Form Item Component */
 export class FormItem extends ElFormItem {}
 
-/** Header Component */
-export class Header extends ElHeader {}
-
 /** Input Component */
 export class Input extends ElInput {}
 
 /** Input Number Component */
 export class InputNumber extends ElInputNumber {}
-
-/** Main Component */
-export class Main extends ElMain {}
 
 /** Menu that provides navigation for your website */
 export class Menu extends ElMenu {}
@@ -231,32 +173,11 @@ export class Option extends ElOption {}
 /** Dropdown Select Option Group Component */
 export class OptionGroup extends ElOptionGroup {}
 
-/** Pagination Component */
-export class Pagination extends ElPagination {}
-
 /** Popover Component */
 export class Popover extends ElPopover {}
 
-/** Progress Component */
-export class Progress extends ElProgress {}
-
 /** Rate Component */
 export class Rate extends ElRate {}
-
-/** Radio Component */
-export class Radio extends ElRadio {}
-
-/** Radio Button Component */
-export class RadioButton extends ElRadioButton {}
-
-/** Radio Group Component */
-export class RadioGroup extends ElRadioGroup {}
-
-/** Row Layout Component */
-export class Row extends ElRow {}
-
-/** Dropdown Select Component */
-export class Select extends ElSelect {}
 
 /** Slider Component */
 export class Slider extends ElSlider {}
@@ -270,9 +191,6 @@ export class Steps extends ElSteps {}
 /** Submenu Component */
 export class Submenu extends ElSubmenu {}
 
-/** Switch Component */
-export class Switch extends ElSwitch {}
-
 /** Table Component */
 export class Table extends ElTable {}
 
@@ -284,9 +202,6 @@ export class Tabs extends ElTabs {}
 
 /** Tab Pane Component */
 export class TabPane extends ElTabPane {}
-
-/** Tag Component */
-export class Tag extends ElTag {}
 
 /** Timeline Component */
 export class Timeline extends ElTimeline {}
@@ -315,9 +230,6 @@ export class Upload extends ElUpload {}
 /** Divider Component */
 export class Divider extends ElDivider {}
 
-/** Link Component */
-export class Link extends ElLink {}
-
 /** Image Component */
 export class Image extends ElImage {}
 
@@ -331,13 +243,10 @@ export class Calendar extends ElCalendar {}
 export class Backtop extends ElBacktop {}
 
 /** InfiniteScroll Directive */
-export const InfiniteScroll: PluginObject<ElInfiniteScroll>;
+export const InfiniteScroll: PluginObject<ElInfiniteScroll>
 
 /** PageHeader Component */
 export class PageHeader extends ElPageHeader {}
-
-/** Avatar Component */
-export class Avatar extends ElAvatar {}
 
 /** Drawer Component */
 export class Drawer extends ElDrawer {}