+++ CHANGELOG.md
@@ -1,3 +1,80 @@
+## [0.0.24](https://github.com/kkbjs/element3/compare/v0.0.23...v0.0.24) (2020-11-12)
+
+
+
+## [0.0.23](https://github.com/kkbjs/element3/compare/v0.0.22...v0.0.23) (2020-11-12)
+
+
+### Bug Fixes
+
+* fix element3.d.ts ([05b0e60](https://github.com/kkbjs/element3/commit/05b0e6015d0bac7610d716a58dd84bd21114e457))
+* Radio.vue and RadioButton.vue get elForm disabled attribute error ([ae29201](https://github.com/kkbjs/element3/commit/ae292019620195f880db90b7953845c89312cce5))
+* Radio.vue getting elFormItemSize error ([e778c71](https://github.com/kkbjs/element3/commit/e778c7168a5a1b08d7d81b91aadca85a7435867f))
+* RadioGroup.vue get injection value error ([1f6696c](https://github.com/kkbjs/element3/commit/1f6696c8a0db511ff32a4177b32c749ff3184ba4))
+* refine user default language zh-CN ([2adda01](https://github.com/kkbjs/element3/commit/2adda01e825d5e8358187039d6474144c0f94f7d))
+* refine user default language zh-CN ([4fa3602](https://github.com/kkbjs/element3/commit/4fa3602fb0d1c4ad00479dfeb1e15f6efb7e6442))
+* update Avatar and test ([417b505](https://github.com/kkbjs/element3/commit/417b505cbfac4c6ed59a16ea1f7d5bb63de8d675))
+* update input.md ([f0493e3](https://github.com/kkbjs/element3/commit/f0493e3c0e59ab81a6d15016980908e7af95b8ba))
+* update link and test ([bf4db91](https://github.com/kkbjs/element3/commit/bf4db919f0dd76834fe6b3c2e76c8e9355f69371))
+* update Notification and test ([0045efb](https://github.com/kkbjs/element3/commit/0045efb68e9295ce5ce8d34b9773b49804d23607))
+* update rate and check ([7afa77f](https://github.com/kkbjs/element3/commit/7afa77f3fb3f53db589b31a17693c94e7000a0ff))
+* user language default zh-CN ([d5145cb](https://github.com/kkbjs/element3/commit/d5145cb4a1d93f199224b6cbb78deb3fec35eaac))
+
+
+### Features
+
+*  check button and breadcrumb and  NavMenu  and Tabs component ([ea59038](https://github.com/kkbjs/element3/commit/ea590386ce7868871b8386b9c5ea17612bd90be1))
+* change messagebox api ([034f432](https://github.com/kkbjs/element3/commit/034f43207fbc3375783214fbe60e6e964b978aac))
+* messageBox api add $ prefix ([a558746](https://github.com/kkbjs/element3/commit/a5587468afd72b8ff39da91e2caa133e29a6ec67))
+
+
+
+## [0.0.22](https://github.com/kkbjs/element3/compare/v0.0.21...v0.0.22) (2020-11-08)
+
+
+### Features
+
+* add setupGlobalOptions api for set global options ([26ad5cb](https://github.com/kkbjs/element3/commit/26ad5cba75a216d6adef8520aaa92ec55efcadcf))
+* change useMsgbox api ([2b55fc8](https://github.com/kkbjs/element3/commit/2b55fc88d31330a3621ac9ab3b5fe57243707872))
+
+
+
+## [0.0.21](https://github.com/kkbjs/element3/compare/v0.0.20...v0.0.21) (2020-11-07)
+
+
+### Bug Fixes
+
+* update doc and test ([8c71671](https://github.com/kkbjs/element3/commit/8c7167127b9d1a327e66355752622263607365bc))
+
+
+
+## [0.0.20](https://github.com/kkbjs/element3/compare/v0.0.19...v0.0.20) (2020-11-05)
+
+
+### Bug Fixes
+
+* avater ... md ([7fb82a2](https://github.com/kkbjs/element3/commit/7fb82a2cdee69c7f480eb1d12243538c83c517e5))
+* cascader export path ([1db3ca7](https://github.com/kkbjs/element3/commit/1db3ca725f5e182f98b3fb150b08e627fd7b42ae))
+
+
+
+## [0.0.19](https://github.com/kkbjs/element3/compare/v0.0.18...v0.0.19) (2020-11-03)
+
+
+### Bug Fixes
+
+* change emits update:value to update:modelValue ([a957075](https://github.com/kkbjs/element3/commit/a95707549f2cb0f06a12bd2ea383ab65219ad05f))
+* change installation.md & deleted test file ([3b1c3ca](https://github.com/kkbjs/element3/commit/3b1c3ca494b93c68557a05a42d5aad492fd16996))
+* change slot to vue3 version ([0b4dcf1](https://github.com/kkbjs/element3/commit/0b4dcf168203868b4314fe25fc7941c0ca8327cb))
+* convert all absolute paths to relative paths ([aec723d](https://github.com/kkbjs/element3/commit/aec723df0b69af04631f0fcb3eb51f209d80c09b))
+* export carousel-item & add some comp in entry.js ([55b3a2a](https://github.com/kkbjs/element3/commit/55b3a2a4e76bff17bfc2e1a3e5b4771f74629939))
+* export option component ([33d676a](https://github.com/kkbjs/element3/commit/33d676a4ccb9fd9df8a2beec619e6670e5fe8765))
+* expose update in scrollbar.js to fix a error ([f83b723](https://github.com/kkbjs/element3/commit/f83b72376f886fa19c572ab98a892f337caac4be))
+* fix addResizeListener params wrong ([89f40b8](https://github.com/kkbjs/element3/commit/89f40b8f306ea3cdabc1497cb86b5dbbbc387fcd))
+* fix avatar error callback call way ([4ff9e33](https://github.com/kkbjs/element3/commit/4ff9e33465f67f304ec988e74097ae9ff4b43c34))
+
+
+
 ## [0.0.18](https://github.com/kkbjs/element3/compare/v0.0.17...v0.0.18) (2020-10-27)